UX: Consistently prefer Uint8Array less conversion to Int16Array for 50% speedup

+++ src/write/toSoundFont2.ts
@@ -68,7 +68,7 @@ export const toSoundFont2 = async (
   rmdirSync(folderPath)
 
   soundFont.metaData.version = '2.04'
-  soundFont.sampleData = new Int16Array(sampleBuffer)
+  soundFont.sampleData = new Uint8Array(sampleBuffer)
   soundFont.presetData.sampleHeaders = sampleHeaders
   return new SoundFont3(new Uint8Array(writeSoundFont(soundFont)))
 }
